refactor: reorg integration tests
