+++ packages/sample/package.json
@@ -12,8 +12,7 @@
     "build": "rm -rf lib/ && tsc",
     "start": "node lib/index.js",
     "dev": "pnpm build && pnpm start",
-    "lint": "eslint src/",
-    "test:integration": "node --test"
+    "lint": "eslint src/"
   },
   "devDependencies": {
     "@silverhand/eslint-config": "^1.3.0",
@@ -29,7 +28,6 @@
   },
   "prettier": "@silverhand/eslint-config/.prettierrc",
   "dependencies": {
-    "@faker-js/faker": "^7.6.0",
     "@silverhand/essentials": "^1.3.0",
     "@withtyped/client": "workspace:*",
     "@withtyped/postgres": "workspace:*",
